Code Refactor
Create a function that returns a Column for the Subway Line, to check for use case where data in array changes
Create files for ViewModel Items and remove from Subway Result Model.
Refactor viewForItem on TableView
Run FastLane
