+++ RipRoadNYCSubway/Features/Model/ViewModel/SubwayLineViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  SubwayLineViewModel.swift
+//  RipRoadNYCSubway
+//
+//  Created by Daniel Spady on 2021-04-06.
+//
+
+import Foundation
+
+struct SubwayListViewModel {
+    var isSelected: Bool = false
+    var itemValue: String = ""
+}
