+++ fastlane/report.xml
@@ -5,17 +5,17 @@
     
     
       
-      <testcase classname="fastlane.lanes" name="0: Verifying fastlane version" time="0.000473">
+      <testcase classname="fastlane.lanes" name="0: Verifying fastlane version" time="0.00181">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="1: default_platform" time="0.000175">
+      <testcase classname="fastlane.lanes" name="1: default_platform" time="0.00021">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="2: scan" time="35.802467">
+      <testcase classname="fastlane.lanes" name="2: scan" time="37.97755">
         
       </testcase>
