+++ fastlane/test_output/report.html
@@ -121,6 +121,40 @@
     <section id="test-suites">
       
         
+        <section class="test-suite passing" id="RipRoadNYCSubwayUITests.RipRoadNYCSubwayUITests">
+          <section class="heading" onclick="toggleTests(this);">
+            <h3 class="title">RipRoadNYCSubwayUITests.RipRoadNYCSubwayUITests</h3>
+          </section>
+          <section class="tests">
+            <table>
+            
+              
+              <tr class="test passing "  onclick="toggleDetails('testExample');">
+                <td>
+                  
+                    <h3 class="time">2.279s</h3>
+                  
+                </td>
+                <td><h3 class="title">testExample</h3></td>
+              </tr>
+              
+            
+              
+              <tr class="test passing odd"  onclick="toggleDetails('testLaunchPerformance');">
+                <td>
+                  
+                    <h3 class="time">16.512s</h3>
+                  
+                </td>
+                <td><h3 class="title">testLaunchPerformance</h3></td>
+              </tr>
+              
+            
+            </table>
+          </section>
+        </section>
+      
+        
         <section class="test-suite passing" id="RipRoadNYCSubwayTests.RipRoadNYCSubwayTests">
           <section class="heading" onclick="toggleTests(this);">
             <h3 class="title">RipRoadNYCSubwayTests.RipRoadNYCSubwayTests</h3>
@@ -132,7 +166,7 @@
               <tr class="test passing "  onclick="toggleDetails('testExample');">
                 <td>
                   
-                    <h3 class="time">0.003s</h3>
+                    <h3 class="time">0.005s</h3>
                   
                 </td>
                 <td><h3 class="title">testExample</h3></td>
@@ -143,7 +177,7 @@
               <tr class="test passing odd"  onclick="toggleDetails('testPerformanceExample');">
                 <td>
                   
-                    <h3 class="time">0.358s</h3>
+                    <h3 class="time">0.596s</h3>
                   
                 </td>
                 <td><h3 class="title">testPerformanceExample</h3></td>
@@ -166,7 +200,7 @@
               <tr class="test passing "  onclick="toggleDetails('testBaseResponse');">
                 <td>
                   
-                    <h3 class="time">0.007s</h3>
+                    <h3 class="time">0.005s</h3>
                   
                 </td>
                 <td><h3 class="title">testBaseResponse</h3></td>
@@ -177,7 +211,7 @@
               <tr class="test passing odd"  onclick="toggleDetails('testBaseService');">
                 <td>
                   
-                    <h3 class="time">0.108s</h3>
+                    <h3 class="time">0.112s</h3>
                   
                 </td>
                 <td><h3 class="title">testBaseService</h3></td>
@@ -188,7 +222,7 @@
               <tr class="test passing "  onclick="toggleDetails('testBaseServiceHandlesErrors');">
                 <td>
                   
-                    <h3 class="time">0.007s</h3>
+                    <h3 class="time">0.024s</h3>
                   
                 </td>
                 <td><h3 class="title">testBaseServiceHandlesErrors</h3></td>
@@ -234,7 +268,7 @@
               <tr class="test passing "  onclick="toggleDetails('testSubwayLineViewModel');">
                 <td>
                   
-                    <h3 class="time">0.041s</h3>
+                    <h3 class="time">0.048s</h3>
                   
                 </td>
                 <td><h3 class="title">testSubwayLineViewModel</h3></td>
@@ -245,7 +279,7 @@
               <tr class="test passing odd"  onclick="toggleDetails('testSubwayLineViewModelFilter');">
                 <td>
                   
-                    <h3 class="time">0.036s</h3>
+                    <h3 class="time">0.033s</h3>
                   
                 </td>
                 <td><h3 class="title">testSubwayLineViewModelFilter</h3></td>
@@ -256,7 +290,7 @@
               <tr class="test passing "  onclick="toggleDetails('testSubwayModel');">
                 <td>
                   
-                    <h3 class="time">0.079s</h3>
+                    <h3 class="time">0.064s</h3>
                   
                 </td>
                 <td><h3 class="title">testSubwayModel</h3></td>
@@ -267,7 +301,7 @@
               <tr class="test passing odd"  onclick="toggleDetails('testSubwayModelNumberOfItems');">
                 <td>
                   
-                    <h3 class="time">0.014s</h3>
+                    <h3 class="time">0.015s</h3>
                   
                 </td>
                 <td><h3 class="title">testSubwayModelNumberOfItems</h3></td>
@@ -278,7 +312,7 @@
               <tr class="test passing "  onclick="toggleDetails('testSubwayViewModelItem');">
                 <td>
                   
-                    <h3 class="time">0.012s</h3>
+                    <h3 class="time">0.014s</h3>
                   
                 </td>
                 <td><h3 class="title">testSubwayViewModelItem</h3></td>
@@ -289,44 +323,10 @@
               <tr class="test passing odd"  onclick="toggleDetails('testSubwyModelIsSuccessful');">
                 <td>
                   
-                    <h3 class="time">0.012s</h3>
+                    <h3 class="time">0.013s</h3>
                   
                 </td>
                 <td><h3 class="title">testSubwyModelIsSuccessful</h3></td>
-              </tr>
-              
-            
-            </table>
-          </section>
-        </section>
-      
-        
-        <section class="test-suite passing" id="RipRoadNYCSubwayUITests.RipRoadNYCSubwayUITests">
-          <section class="heading" onclick="toggleTests(this);">
-            <h3 class="title">RipRoadNYCSubwayUITests.RipRoadNYCSubwayUITests</h3>
-          </section>
-          <section class="tests">
-            <table>
-            
-              
-              <tr class="test passing "  onclick="toggleDetails('testExample');">
-                <td>
-                  
-                    <h3 class="time">1.886s</h3>
-                  
-                </td>
-                <td><h3 class="title">testExample</h3></td>
-              </tr>
-              
-            
-              
-              <tr class="test passing odd"  onclick="toggleDetails('testLaunchPerformance');">
-                <td>
-                  
-                    <h3 class="time">16.319s</h3>
-                  
-                </td>
-                <td><h3 class="title">testLaunchPerformance</h3></td>
               </tr>
